requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
Babel==2.9.1
boto3==1.17.93
botocore==1.20.93
dj-database-url==0.5.0
Django==3.2.4
django-crispy-forms==1.12.0
django-money==2.0
django-session-timeout==0.1.0
django-storages==1.11.1
docutils==0.17.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
jmespath==0.10.0
Pillow==8.2.0
psycopg2-binary==2.8.6
py-moneyed==1.2
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
reportlab==3.5.67
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
